refacto: refactoring forms and adding validation from API
