+++ src/components/moderatorRessources/thematiquesForm/ThematiquesForm.test.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event';
 import { ThematiquesForm } from './ThematiquesForm';
 import MockAdapter from 'axios-mock-adapter';
 import { axiosInstance } from '@/RequestInterceptor';
-import { moderatorRessources } from '../../../../utils/tests/moderatorRessources.fixtures';
+import { moderatorRessources } from '../../../utils/tests/moderatorRessources.fixtures';
 import { axe, toHaveNoViolations } from 'jest-axe';
 import { ocWelcomeAPIResponse } from '@/utils/tests/ocWelcome.fixtures';
